--- src/features/auth/signup/PhoneValidation/PhoneValidationTooManySMSSent/index.tsx
@@ -1,1 +1,0 @@
-export * from './PhoneValidationTooManySMSSent'
