+++ src/features/identityCheck/pages/phoneValidation/errors/index.tsx
@@ -1,1 +1,2 @@
 export * from './PhoneValidationTooManyAttempts'
+export * from './PhoneValidationTooManySMSSent'
