+++ src/features/identityCheck/pages/phoneValidation/errors/tests/PhoneValidationTooManySMSSent.native.test.tsx
@@ -3,8 +3,8 @@ import React from 'react'
 import { mocked } from 'ts-jest/utils'
 
 import { navigate } from '__mocks__/@react-navigation/native'
-import { PhoneValidationTooManySMSSent } from 'features/auth/signup/PhoneValidation/PhoneValidationTooManySMSSent'
 import { usePhoneValidationRemainingAttempts } from 'features/identityCheck/api/api'
+import { PhoneValidationTooManySMSSent } from 'features/identityCheck/pages/phoneValidation/errors/PhoneValidationTooManySMSSent'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { fireEvent, render } from 'tests/utils'
